move CoreBuilder in arduino/builder and move CompilationDatabase in compilation pkg
